Removed context-path for comments-webservice
If we add context-root, then we cannot mock the server on the comments-webservice side and we will have to make calls with <code>testMode = 'EXPLICIT'</code> for the producer tests on the comments-webservice side. This doesn't work on the comments-webservice side since there is no mock OAuth2 token that gets injected there.